// autogpt_platform/frontend/src/app/api/__generated__/models/executionStartedResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { ExecutionStartedResponseSessionId } from "./executionStartedResponseSessionId";
import type { ExecutionStartedResponseLibraryAgentId } from "./executionStartedResponseLibraryAgentId";
import type { ExecutionStartedResponseLibraryAgentLink } from "./executionStartedResponseLibraryAgentLink";

/**
 * Response for run/schedule actions.
 */
export interface ExecutionStartedResponse {
  type?: ResponseType;
  message: string;
  session_id?: ExecutionStartedResponseSessionId;
  execution_id: string;
  graph_id: string;
  graph_name: string;
  library_agent_id?: ExecutionStartedResponseLibraryAgentId;
  library_agent_link?: ExecutionStartedResponseLibraryAgentLink;
  status?: string;
}